Move HammerPointStates to States.cs
