+++ nekoyume/Assets/_Scripts/BlockChain/ActionRenderHandler.cs
@@ -621,7 +621,7 @@ namespace Nekoyume.BlockChain
                 UpdateAgentStateAsync(eval).Forget();
                 UpdateCurrentAvatarStateAsync(eval).Forget();
                 RenderQuest(avatarAddress, avatarState.questList?.completedQuestIds);
-                RxProps.UpdateHammerPointStates(result.recipeId, hammerPointState);
+                States.UpdateHammerPointStates(result.recipeId, hammerPointState);
 
                 if (!(nextQuest is null))
                 {
@@ -1638,7 +1638,7 @@ namespace Nekoyume.BlockChain
             recipeIds.AddRange(sharedModel.UnlockedRecipes.Value);
             sharedModel.SetUnlockedRecipes(recipeIds);
             sharedModel.UpdateUnlockableRecipes();
-            RxProps.UpdateHammerPointStates(eval.Action.RecipeIds).Forget();
+            States.UpdateHammerPointStates(eval.Action.RecipeIds).Forget();
         }
 
         private void ResponseUnlockWorld(ActionBase.ActionEvaluation<UnlockWorld> eval)
