--- nekoyume/Assets/_Scripts/State/RxProps.Craft.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 860e5920b57d42b499cc987f12bc06f1
-timeCreated: 1660662165
