+++ nekoyume/Assets/_Scripts/State/RxProps.cs
@@ -98,7 +98,6 @@ namespace Nekoyume.State
             }
 
             _currentAvatarAddr = avatarAddr;
-            _hammerPointStates = null;
             OnAvatarChangedArena();
             OnAvatarChangedEvent();
         }
